# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: ketuk-api
  version: 0.2.0
  description: |
    Backend Ketuk. Sukses JSON `{"data": ...}` kecuali auth (`data` + `tokens`) dan callback Duitku (teks `SUCCESS`).
    Schema katalog + undangan + planner sesuai handler hidup. CRUD lokasi/tamu/buku ucapan sudah dikode sesuai `docs/contracts/undangan-crud.md`.
    PATCH inviter dan caption galeri juga sudah dikode sesuai `docs/contracts/undangan-patch-gallery-inviter.md`.
    CRUD planner (patch/archive, hari, kategori, item budget, checklist) sudah dikode sesuai `docs/contracts/planner-crud.md` bagian 1; rundown sesi jam+PIC belum.
    Auth dan billing masih path-only. Vendor belum diimplementasi.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/templates:
    get:
      tags: [Catalog]
      operationId: listTemplates
      summary: Template undangan
      description: "Publik. Bearer opsional — jika JWT valid dan paket punya `premium_templates`, template premium `locked=false`."
      parameters:
        - in: query
          name: type
          description: "Filter slug `event_types` (wedding, birthday, …)"
          schema: {type: string}
      responses:
        "200":
          description: Daftar template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/invitations:
    get:
      tags: [Invitation]
      operationId: listInvitations
      summary: Daftar undangan non-archived milik host
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: "List. Slice kosong di Go bisa `data: null`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: createInvitation
      summary: Buat undangan draft
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "201":
          description: "Draft baru. `type_id` diambil dari template. `content` kosong jadi `{}`."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/EntitlementDenied"
        "404":
          description: template tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: getInvitation
      summary: Detail undangan
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Invitation]
      operationId: patchInvitation
      summary: Ubah judul, content, atau tanggal
      description: "Field yang tidak dikirim tidak diubah. `content` hanya diubah jika key JSON hadir (bukan null)."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchInvitationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: archiveInvitation
      summary: Archive undangan (bukan hard delete)
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Diarsipkan
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/publish:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    post:
      tags: [Invitation]
      operationId: publishInvitation
      summary: Publish dan isi slug
      description: "Slug bukan jalur RSVP. Idempoten jika sudah published. Tamu memakai `/v1/public/rsvp/{token}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/invitations/{id}/locations:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listLocations
      summary: Daftar lokasi
      description: "Ubah / hapus satu lokasi: `/v1/invitations/{id}/locations/{locationId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addLocation
      summary: Tambah lokasi + maps_url
      description: "Tempel URL peta. Server tidak geocode. `id` di body diabaikan."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Location"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/locations/{locationId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/LocationId"
    patch:
      tags: [Invitation]
      operationId: patchLocation
      summary: Ubah lokasi
      description: "Body parsial: field yang tidak dikirim tidak diubah. `id` di body diabaikan. Server tidak geocode `maps_url`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchLocationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteLocation
      summary: Hapus lokasi
      description: "Lokasi bukan milik undangan ini → 404 `location`."
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/guests:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listGuests
      summary: Daftar tamu + rsvp_token
      description: "Ubah / hapus satu tamu: `/v1/invitations/{id}/guests/{guestId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addGuest
      summary: Tambah tamu
      description: "Status awal `draft`. `rsvp_token` UUID baru untuk `/v1/public/rsvp/{token}`."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/EntitlementDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/guests/{guestId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/GuestId"
    patch:
      tags: [Invitation]
      operationId: patchGuest
      summary: Ubah tamu atau tandai invited
      description: "Body parsial. `status` host hanya `draft`|`invited` — jawaban RSVP tamu tidak lewat sini. `rsvp_token` tetap. Tidak mengirim notifikasi."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchGuestRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteGuest
      summary: Hapus tamu
      description: "Tautan RSVP tamu ikut mati. Kuota tamu menghitung baris, jadi slot kembali terpakai."
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    InvitationId:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
    LocationId:
      name: locationId
      in: path
      required: true
      schema: {type: string, format: uuid}
    GuestId:
      name: guestId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    BadRequest:
      description: "JSON rusak (`invalid_json`) atau validasi (`invalid_input` / `invalid_id`)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InvalidId:
      description: Path UUID rusak
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: invalid_id, message: invalid id}
    Unauthorized:
      description: JWT hilang atau tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: unauthorized, message: missing bearer token}
    EntitlementDenied:
      description: Kuota atau template premium
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          examples:
            invitationQuota:
              value:
                error: {code: entitlement_denied, message: max_active_invitations quota exceeded}
            guestQuota:
              value:
                error: {code: entitlement_denied, message: max_guests_per_invitation quota exceeded}
            premiumTemplate:
              value:
                error: {code: entitlement_denied, message: premium_templates not included in current plan}
            plannerQuota:
              value:
                error: {code: entitlement_denied, message: max_active_planners quota exceeded}
            plannerDaily:
              value:
                error: {code: entitlement_denied, message: planner_daily not included in current plan}
    NotFound:
      description: Resource tidak ada atau bukan milik host; RSVP token/slug unpublished
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          examples:
            invitation:
              value:
                error: {code: not_found, message: invitation not found}
            plannerChild:
              summary: "Anak planner: day / session / category / budget_item / checklist"
              value:
                error: {code: not_found, message: budget_item not found}
...
    InternalError:
      description: Kesalahan server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: internal, message: internal server error}
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
...
    Template:
      type: object
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        type_id: {type: string, format: uuid}
        name: {type: string}
        tier:
          type: string
          enum: [free, premium]
        version: {type: integer}
        slots:
          description: JSON array nama slot template
        locked:
          type: boolean
          description: true jika premium dan caller tidak berhak
    Invitation:
      type: object
      properties:
        id: {type: string, format: uuid}
        owner_id: {type: string, format: uuid}
        type_id: {type: string, format: uuid}
        template_id: {type: string, format: uuid}
        template_version: {type: integer}
        slug:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, published, archived]
        title: {type: string}
        content:
          type: object
          additionalProperties: true
          description: Objek bebas. Key mengikuti invitation_schema tipe acara; server tidak validasi.
        starts_at:
          type: string
          format: date-time
          nullable: true
        ends_at:
          type: string
          format: date-time
          nullable: true
    Location:
      type: object
      properties:
        id: {type: string, format: uuid}
        label: {type: string}
        venue_name: {type: string}
        address: {type: string}
        maps_url: {type: string, description: URL Google/Apple Maps yang ditempel}
        lat:
          type: number
          format: double
          nullable: true
          description: Tidak diisi otomatis
        lng:
          type: number
          format: double
          nullable: true
        starts_at:
          type: string
          format: date-time
          nullable: true
        sort_order: {type: integer}
    Guest:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        group: {type: string}
        rsvp_token: {type: string}
        status:
          type: string
          enum: [draft, invited, opened, accepted, declined]
        plus_ones: {type: integer}
        message: {type: string, description: Pesan RSVP, bukan buku ucapan}
        rsvp_at:
          type: string
          format: date-time
          nullable: true
...
    CreateInvitationRequest:
      type: object
      required: [title, template_id]
      properties:
        title: {type: string}
        template_id: {type: string, format: uuid}
        content:
          type: object
          additionalProperties: true
    PatchInvitationRequest:
      type: object
      properties:
        title: {type: string}
        content:
          type: object
          additionalProperties: true
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
...
    CreateGuestRequest:
      type: object
      required: [name]
      properties:
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        group: {type: string}
    PatchLocationRequest:
      type: object
      description: Semua field opsional. Jangan geocode maps_url.
      properties:
        label: {type: string}
        venue_name: {type: string}
        address: {type: string}
        maps_url: {type: string}
        lat: {type: number, format: double, nullable: true}
        lng: {type: number, format: double, nullable: true}
        starts_at: {type: string, format: date-time, nullable: true}
        sort_order: {type: integer}
    PatchGuestRequest:
      type: object
      description: "Semua opsional. status host hanya draft|invited. Jangan kirim rsvp_token."
      properties:
        name: {type: string}
        email: {type: string}
        phone: {type: string}
        group: {type: string}
        status:
          type: string
          enum: [draft, invited]
...
    TemplateListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Template"}
          nullable: true
    InvitationResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/Invitation"}
    InvitationListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Invitation"}
          nullable: true
    LocationResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/Location"}
    LocationListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Location"}
          nullable: true
    GuestResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/Guest"}
    GuestListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Guest"}
          nullable: true
...